# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Zense API Docs",
        "title": "Zense",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/forums": {
            "get": {
                "description": "Get all forum posts",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Forums"
                ],
                "summary": "Get All Forums",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/web.ForumResponse"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a new forum post",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Forums"
                ],
                "summary": "Create Forum",
                "parameters": [
                    {
                        "description": "Forum Data",
                        "name": "forum",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/web.ForumCreate"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/web.ForumResponse"
                        }
                    }
                }
            }
        },
        "/forums/{id}": {
            "get": {
                "description": "Get a forum post by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Forums"
                ],
                "summary": "Get Forum by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Forum ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/web.ForumResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update a forum post",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Forums"
                ],
                "summary": "Update Forum",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Forum ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated Forum Data",
                        "name": "forum",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/web.ForumUpdate"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/web.ForumResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete a forum post",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Forums"
                ],
                "summary": "Delete Forum",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Forum ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
        "/forums/{id}/topic": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Remove a topic from a specific forum by its ID",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Forums"
                ],
                "summary": "Remove Topic from Forum",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Forum ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Remove Topic from Forum",
                        "name": "topic",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/web.ForumRemoveTopic"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
        "/journals": {
            "get": {
                "description": "Get all journals",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Journals"
                ],
                "summary": "Get All Journals",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/web.JournalResponse"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a new journal entry",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Journals"
                ],
                "summary": "Create Journal",
                "parameters": [
                    {
                        "description": "Journal Data",
                        "name": "journal",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/web.JournalCreate"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/web.JournalResponse"
                        }
                    }
                }
            }
        },
        "/journals/{id}": {
            "get": {
                "description": "Get journal by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Journals"
                ],
                "summary": "Get Journal by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Journal ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/web.JournalResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update an existing journal entry",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Journals"
                ],
                "summary": "Update Journal",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Journal ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated Journal Data",
                        "name": "journal",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/web.JournalUpdate"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/web.JournalResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete a journal entry",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Journals"
                ],
                "summary": "Delete Journal",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Journal ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "domain.JournalMood": {
            "type": "string",
            "enum": [
                "happy",
                "good",
                "normal",
                "sad",
                "angry"
            ],
            "x-enum-varnames": [
                "Happy",
                "Good",
                "Normal",
                "Sad",
                "Angry"
            ]
        },
        "domain.JournalVisibility": {
            "type": "string",
            "enum": [
                "private",
                "public"
            ],
            "x-enum-varnames": [
                "PrivateJournal",
                "PublicJournal"
            ]
        },
...
        "web.ForumCreate": {
            "type": "object",
            "required": [
                "content",
                "title",
                "topics"
            ],
            "properties": {
                "content": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "topics": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "web.ForumRemoveTopic": {
            "type": "object",
            "required": [
                "topic_id"
            ],
            "properties": {
                "id": {
                    "type": "integer"
                },
                "topic_id": {
                    "type": "integer"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "web.ForumResponse": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                },
                "topics": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/web.TopicResponse"
                    }
                },
                "updated_at": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/web.UserResponse"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "web.ForumUpdate": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                },
                "topics": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "web.JournalCreate": {
            "type": "object",
            "required": [
                "content"
            ],
            "properties": {
                "content": {
                    "type": "string"
                },
                "mood": {
                    "$ref": "#/definitions/domain.JournalMood"
                },
                "user_id": {
                    "type": "integer"
                },
                "visibility": {
                    "$ref": "#/definitions/domain.JournalVisibility"
                }
            }
        },
        "web.JournalResponse": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "mood": {
                    "$ref": "#/definitions/domain.JournalMood"
                },
                "updated_at": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/web.UserResponse"
                },
                "user_id": {
                    "type": "integer"
                },
                "visibility": {
                    "$ref": "#/definitions/domain.JournalVisibility"
                }
            }
        },
        "web.JournalUpdate": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "mood": {
                    "enum": [
                        "happy",
                        "good",
                        "normal",
                        "sad",
                        "angry"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/domain.JournalMood"
                        }
                    ]
                },
                "user_id": {
                    "type": "integer"
                },
                "visibility": {
                    "enum": [
                        "private",
                        "public"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/domain.JournalVisibility"
                        }
                    ]
                }
            }
        },
...
        "web.TopicResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "web.UserResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Provide your JWT token here. Example: \"Bearer {token}\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}